BookStore/WebApi/Application/BookOperations/Commands/CreateBook/CreateBookCommand.cs
using AutoMapper;
using WebApi.Common;
using WebApi.DBOperations;
using WebApi.Entities;

namespace WebApi.Application.BookOperations.Commands.CreateBook
{
    public class CreateBookCommand
    {
        private readonly IBookStoreDbContext _context;
        private readonly IMapper _mapper;
        public CreateBookModel Model { get; set; }

        public CreateBookCommand(IBookStoreDbContext context, IMapper mapper)
        {
            _context = context;
            _mapper = mapper;
        }

        public void Handle()
        {
            var existingBook = _context.Books.SingleOrDefault(x => x.Title.ToLower() == Model.Title.ToLower().Trim());
            if(existingBook is not null)
                throw new InvalidOperationException("The book is already exist");
            var book = _mapper.Map<Book>(Model);
            _context.Books.Add(book);
            _context.SaveChanges();
        }
    }
    
    public class CreateBookModel
    {
        public string Title { get; set; }        
        public int PageCount { get; set; }
        public DateTime PublishDate { get; set; }
        public int GenreId { get; set; }
        public int AuthorId { get; set; }
    }
}